refactor(api): extract haversine to shared lib (kills last cross-context import)
Finding 4 from the bounded-context audit. haversineDistance was a generic
geo math primitive trapped in matching-utils. scheduling/venue.ts had to
import from matching/matching to use it — the only remaining cross-context
source import after PR #302.

Move:
- haversineDistance: matching-utils.ts → packages/api/src/lib/geo.ts

Update imports:
- matching/matching.ts: from "./matching-utils" → from "../../lib/geo"
- scheduling/venue.ts: from "../matching/matching" → from "../../lib/geo"
- Drop matching/matching.ts re-export of haversineDistance (no longer
  needed)

Verified: zero cross-context source imports remain in packages/api/src/
contexts/. All cross-context coupling now goes through:
  - shared lib (lib/geo)
  - shared schema (db/schema/sip-and-speak barrel)
  - domain events (domain-events.ts)
  - shared infra (auth, db)

Type errors: 6 unchanged (chat.ts pre-existing). Tests unchanged.
