+++ packages/api/src/contexts/matching/matching-utils.ts
@@ -3,23 +3,6 @@
 
 const MAX_RADIUS_KM = 50;
 
-/** Haversine distance in km between two lat/lng points */
-export function haversineDistance(
-  lat1: number,
-  lon1: number,
-  lat2: number,
-  lon2: number,
-): number {
-  const R = 6371;
-  const toRad = (deg: number) => (deg * Math.PI) / 180;
-  const dLat = toRad(lat2 - lat1);
-  const dLon = toRad(lon2 - lon1);
-  const a =
-    Math.sin(dLat / 2) ** 2 +
-    Math.cos(toRad(lat1)) * Math.cos(toRad(lat2)) * Math.sin(dLon / 2) ** 2;
-  return R * 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-}
-
 /**
  * Language complementarity score.
  * 1.0 — partner speaks a language user wants to learn AND partner wants to learn a language user speaks.
